some changes done

+++ Home.js
@@ -1,4 +1,24 @@
+// var mainApp ={};
+
+// (function(){
+//     varfirebase = app_fireBase;
+//     var uid = null
+//     firebase.auth().onAuthStateChanged(function(user) {
+//         if (user) {
+//           // User is signed in.
+//         }else{
+//             // redirect to login page
+//             window.location.replace("login.html");
+//         }
+//       });
+//       function logOut(){
+//           firebase.auth().signOut();
+//       }
+//       mainApp.logOut =logOut;
+// })()
+
 console.log('Welcome to Notes-App');
+
 //javascript is yet to be added
 showNotes();
 let addbtn = document.getElementById('addbtn');
@@ -22,20 +42,20 @@ addbtn.addEventListener("click", function (e) {
 function showNotes() {
     let notes = localStorage.getItem("notes");
     if (notes == null) {
-        notesObj = [];
-    }
-    else {
-        notesObj = JSON.parse(notes);
+      notesObj = [];
+    } else {
+      notesObj = JSON.parse(notes);
     }
     let html = "";
-    notesObj.forEach(function (element, index) {
-        html += `
+    notesObj.forEach(function(element, index) {
+      html += `
             
             <div id ="VNote" class="container">
-      <h5 class="card-title">Note ${index + 1} </h5>
+      <h5 class="card-title">Note No. ${index + 1} </h5>
       <p class="card txt">${element}</p>
-      <button id="deletebtn" class="deletebtn" onclick="deleteNote(this.id)">Delete Note</button>
-      <button id="editbtn" class="editbtn">Edit Note</button>
+      <button  class="deletebtn" onclick="deleteNote(this.id)">Delete Note</button>
+      <button  onclick="editbtn(this.id)" class="editbtn"> Edit Note </button>
+      
     </div>
             `;
     });
@@ -45,27 +65,30 @@ function showNotes() {
 
     }
     else {
-        notesElm.innerHTML = `Nothing to show! Please add note first`
+        notesElm.innerHTML = `Nothing to show! Please add some notes first`
     }
+
 }
-
-
 function deleteNote(index) {
     //   console.log("I am deleting", index);
+    let confirmation = confirm("Are you sure, you want to delete Note?")
+    if (confirmation == true) {
+        let notes = localStorage.getItem("notes");
+        if (notes == null) {
+            notesObj = [];
+        } else {
+            notesObj = JSON.parse(notes);
+        }
 
-    let notes = localStorage.getItem("notes");
-    if (notes == null) {
-        notesObj = [];
-    } else {
-        notesObj = JSON.parse(notes);
+        notesObj.splice(index, 1);
+        localStorage.setItem("notes", JSON.stringify(notesObj));
+        showNotes();
     }
-
-    notesObj.splice(index, 1);
-    localStorage.setItem("notes", JSON.stringify(notesObj));
-    showNotes();
 }
-let editbtn = document.getElementById('editbtn');
-editbtn.addEventListener("click", function (e) {
+
+// let editbtn = document.getElementById('editbtn');
+// editbtn.addEventListener("click", function (e) {
+  function  editbtn(element,index){
     // let Note = document.getElementById("Note");
     let notes = localStorage.getItem("notes");
     if (notes == null) {
@@ -73,36 +96,17 @@ editbtn.addEventListener("click", function (e) {
     } else {
         notesObj = JSON.parse(notes);
     }
-    let txt = ""
-    function edit(element, index) {
-        txt += `<div class="container">
-        <h1>Edit Your note here</h1>
-        <div>
-          <textarea id="Notes" class="Note">${element}</textarea>
-        </div>
-        <button id="addebtn" class="addebtn">Add Edited Note</button>
-      </div>
-        `
-    }
-    edit();
-    let ENote = document.getElementById("eNote");
-    ENote.innerHTML = txt;
+    console.log(notesObj);
+    notesObj.findIndex((element,index) =>{
+        Note.value = element.notes;
+    })
+    notesObj.splice(index,1);
+    localStorage.setItem("notes",JSON.stringify(notesObj));
+showNotes();
+}
+    
 
-    // let addebtn = document.getElementById('addebtn');
-    // addebtn.addEventListener("click", function (e) {
-    //     let Notes = document.getElementById("Notes");
-    //     let notes = localStorage.getItem("notes");
-    //     if (notes == null) {
-    //         notesObj = [];
-    //     }
-    //     else {
-    //         notesObj = JSON.parse(notes);
-    //     }
-    //     notesObj.push(Note.value);
-    //     notesObj.push(Notes.value);
-    //     localStorage.setItem("notes", JSON.stringify(notesObj));
-    //     Notes.value = "";
-    //     console.log(notesObj);
-    //     showNotes();
 
-})
+
+   
+
